PT-20 #time 2d Total work logged #resolve #comment Had to create new webservice calls, and change the screen around a little to get it work correctly. Still might have issues, not that happy with this implementation.

+++ Apromore-Portal/src/main/java/org/apromore/portal/dialogController/similarityclusters/visualisation/ClusterVisualisationController.java
@@ -71,7 +71,7 @@ public class ClusterVisualisationController extends BaseController {
                 if (event.getData() != null) {
                     Object[] idList = (Object[]) event.getData();
                     for (Object anIdList : idList) {
-                        pairwiseFilter.add(((Integer) anIdList).toString());
+                        pairwiseFilter.add((anIdList).toString());
                     }
                     try {
                         Clients.evalJavaScript("clusterVisualisation.refreshData(" + writeUpdatedEdges(pairwiseFilter) + ");");
